remove image after failed upload
Remove images which fail to upload from the document. An image which
fails to upload will not be available on subsequent document loading and
thus should not continue to appear in the document if it fails to
upload. Keeping the image in the document is misleading to the user
(even with the "sorry failed to upload" overlay).

This change removes the image from the document and shows a toast
message instead.